ADD 增加用户邀请码

+++ internal/model/entity/cdkey.go
@@ -1,7 +1,7 @@
 /*
  * @Author: cloudyi.li
  * @Date: 2023-05-10 09:58:37
- * @LastEditTime: 2023-05-15 12:06:28
+ * @LastEditTime: 2023-05-18 13:01:13
  * @LastEditors: cloudyi.li
  * @FilePath: /chatserver-api/internal/model/entity/cdkey.go
  */
@@ -16,7 +16,7 @@ import (
 type CdKey struct {
 	Id        int64                 `gorm:"column:id;primary_key;" json:"id"`
 	CodeKey   string                `gorm:"column:code_key" json:"code_key"`
-	Amount    int                   `gorm:"column:amount" json:"amount"`
+	Amount    float64               `gorm:"column:amount" json:"amount"`
 	CreatedAt jtime.JsonTime        `gorm:"column:created_at" json:"created_at"`
 	UpdatedAt jtime.JsonTime        `gorm:"column:updated_at" json:"updated_at"`
 	DeletedAt jtime.JsonTime        `gorm:"column:deleted_at" json:"deleted_at" `
